Extracted class NamedProperty packages.config: Restsharp is now 104.1
