+++ TechTalk.JiraRestClient/TechTalk.JiraRestClient.csproj
@@ -43,6 +43,7 @@
     <Compile Include="Compatibility.cs" />
     <Compile Include="FieldAttribute.cs" />
     <Compile Include="IssueType.cs" />
+    <Compile Include="NamedProperty.cs" />
     <Compile Include="ServerInfo.cs" />
     <Compile Include="WatchersContainer.cs" />
     <Compile Include="IJiraClient.cs" />
