+++ TechTalk.JiraRestClient/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="RestSharp" version="105.0.1" targetFramework="net40-Client" />
+  <package id="RestSharp" version="104.1" targetFramework="net40-Client" />
 </packages>
